Revert "feat: implement task details modal with support for assignees, checkl…"
