+++ src/__test__/integration/TaskCommentsApi.test.ts
@@ -1,35 +1,5 @@
 const mockGetUser = jest.fn();
-
-const DEFAULT_USERS_DATA = [
-  { id: "owner-1", display_name: "Owner One", avatar_url: null },
-  { id: "member-1", display_name: "Member One", avatar_url: null },
-  { id: "user-2", display_name: "User Two", avatar_url: null },
-];
-
-// A generic chainable mock for Supabase query builder
-const createMockChain = (table?: string) => {
-  const chain: any = {
-    select: jest.fn().mockReturnThis(),
-    eq: jest.fn().mockReturnThis(),
-    in: jest.fn().mockReturnThis(),
-    order: jest.fn().mockReturnThis(),
-    single: jest.fn().mockResolvedValue({ data: null, error: null }),
-    maybeSingle: jest.fn().mockResolvedValue({ data: null, error: null }),
-  };
-
-  if (table === "users") {
-    chain.select.mockReturnValue({
-      in: jest.fn().mockResolvedValue({
-        data: DEFAULT_USERS_DATA,
-        error: null,
-      }),
-    });
-  }
-
-  return chain;
-};
-
-const mockFrom = jest.fn().mockImplementation((table: string) => createMockChain(table));
+const mockFrom = jest.fn();
 
 jest.mock("@/utils/supabase/server", () => ({
   createClient: jest.fn().mockResolvedValue({
@@ -86,18 +56,19 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     });
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: null,
-              error: { message: "Not found" },
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: null,
+                error: { message: "Not found" },
+              }),
             }),
           }),
-        });
+        };
       }
-      return chain;
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("999"));
@@ -113,28 +84,31 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     });
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { column_id: 50 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
       if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: null,
-              error: { message: "Column not found" },
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: null,
+                error: { message: "Column not found" },
+              }),
             }),
           }),
-        });
+        };
       }
-      return chain;
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("1"));
@@ -150,52 +124,59 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     });
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { column_id: 50 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
       if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { board_id: 10 },
-              error: null,
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { board_id: 10 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
       if (table === "boards") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: null,
-                error: null,
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: null,
+                  error: null,
+                }),
               }),
             }),
           }),
-        });
+        };
       }
       if (table === "board_members") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: null,
-                error: null,
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: null,
+                  error: null,
+                }),
               }),
             }),
           }),
-        });
+        };
       }
-      return chain;
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("1"));
@@ -228,50 +209,57 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     ];
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { board_id: 10 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "boards") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: { id: 10 },
+              single: jest.fn().mockReturnValue({
+                data: { column_id: 50 },
                 error: null,
               }),
             }),
           }),
-        });
+        };
       }
-      if (table === "comments") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            order: jest.fn().mockReturnValue({
-              data: mockComments,
-              error: null,
+      if (table === "columns") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { board_id: 10 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
-      return chain;
+      if (table === "boards") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: { id: 10 },
+                  error: null,
+                }),
+              }),
+            }),
+          }),
+        };
+      }
+      if (table === "comments") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              order: jest.fn().mockReturnValue({
+                data: mockComments,
+                error: null,
+              }),
+            }),
+          }),
+        };
+      }
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("1"));
@@ -299,62 +287,71 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     ];
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { column_id: 50 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
       if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { board_id: 10 },
-              error: null,
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { board_id: 10 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
       if (table === "boards") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: null,
-                error: null,
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: null,
+                  error: null,
+                }),
               }),
             }),
           }),
-        });
+        };
       }
       if (table === "board_members") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: { user_id: "member-1" },
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: { user_id: "member-1" },
+                  error: null,
+                }),
+              }),
+            }),
+          }),
+        };
+      }
+      if (table === "comments") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              order: jest.fn().mockReturnValue({
+                data: mockComments,
                 error: null,
               }),
             }),
           }),
-        });
+        };
       }
-      if (table === "comments") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            order: jest.fn().mockReturnValue({
-              data: mockComments,
-              error: null,
-            }),
-          }),
-        });
-      }
-      return chain;
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("5"));
@@ -371,50 +368,57 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     });
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { board_id: 10 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "boards") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: { id: 10 },
+              single: jest.fn().mockReturnValue({
+                data: { column_id: 50 },
                 error: null,
               }),
             }),
           }),
-        });
+        };
       }
-      if (table === "comments") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            order: jest.fn().mockReturnValue({
-              data: null,
-              error: { message: "DB comments error" },
+      if (table === "columns") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { board_id: 10 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
-      return chain;
+      if (table === "boards") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: { id: 10 },
+                  error: null,
+                }),
+              }),
+            }),
+          }),
+        };
+      }
+      if (table === "comments") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              order: jest.fn().mockReturnValue({
+                data: null,
+                error: { message: "DB comments error" },
+              }),
+            }),
+          }),
+        };
+      }
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("1"));
@@ -430,122 +434,62 @@ describe("GET /api/tasks/[taskId]/comments", () => {
     });
 
     mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain(table);
       if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { board_id: 10 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "boards") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
+        return {
+          select: jest.fn().mockReturnValue({
             eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: { id: 10 },
+              single: jest.fn().mockReturnValue({
+                data: { column_id: 50 },
                 error: null,
               }),
             }),
           }),
-        });
+        };
       }
-      if (table === "comments") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            order: jest.fn().mockReturnValue({
-              data: [],
-              error: null,
+      if (table === "columns") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              single: jest.fn().mockReturnValue({
+                data: { board_id: 10 },
+                error: null,
+              }),
             }),
           }),
-        });
+        };
       }
-      return chain;
+      if (table === "boards") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              eq: jest.fn().mockReturnValue({
+                maybeSingle: jest.fn().mockReturnValue({
+                  data: { id: 10 },
+                  error: null,
+                }),
+              }),
+            }),
+          }),
+        };
+      }
+      if (table === "comments") {
+        return {
+          select: jest.fn().mockReturnValue({
+            eq: jest.fn().mockReturnValue({
+              order: jest.fn().mockReturnValue({
+                data: [],
+                error: null,
+              }),
+            }),
+          }),
+        };
+      }
+      return {};
     });
 
     const res = await GET(makeRequest(), makeParams("1"));
     expect(res.status).toBe(200);
     const body = await res.json();
     expect(body).toEqual([]);
-  });
-
-  // ── 500: Users fetch fails ──────────────────────────────
-  it("returns 500 when users data fetch fails", async () => {
-    mockGetUser.mockResolvedValue({
-      data: { user: { id: "owner-1" } },
-    });
-
-    mockFrom.mockImplementation((table: string) => {
-      const chain = createMockChain();
-      if (table === "tasks") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { column_id: 50 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "columns") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            single: jest.fn().mockReturnValue({
-              data: { board_id: 10 },
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "boards") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            eq: jest.fn().mockReturnValue({
-              maybeSingle: jest.fn().mockReturnValue({
-                data: { id: 10 },
-                error: null,
-              }),
-            }),
-          }),
-        });
-      }
-      if (table === "comments") {
-        chain.select.mockReturnValue({
-          eq: jest.fn().mockReturnValue({
-            order: jest.fn().mockReturnValue({
-              data: [{ id: 1, user_id: "owner-1", content: "Test" }],
-              error: null,
-            }),
-          }),
-        });
-      }
-      if (table === "users") {
-        chain.select.mockReturnValue({
-          in: jest.fn().mockResolvedValue({
-            data: null,
-            error: { message: "Users DB error" },
-          }),
-        });
-      }
-      return chain;
-    });
-
-    const res = await GET(makeRequest(), makeParams("1"));
-    expect(res.status).toBe(500);
-    const body = await res.json();
-    expect(body.error).toBe("Failed to load comment authors.");
   });
 });
